refactor(auth-api): rename notifications.ts and ports.ts (#1195)
- notifications.ts → notificationTypes.ts (P2: types-only)
- ports.ts → types.ts (P2: types-only)
- Re-add P2 primary-export-match skip lost in squash merge

Rollback: git revert <hash> --no-edit
